Move schemas to top level app folder
