+++ app/api/api_v1/routers/documents.py
@@ -4,15 +4,15 @@ from typing import Union
 
 from fastapi import APIRouter, Depends, HTTPException
 
-from app.api.api_v1.schemas.document import (
-    FamilyAndDocumentsResponse,
-    FamilyDocumentWithContextResponse,
-)
 from app.clients.db.session import get_db
 from app.repository.document import (
     get_family_and_documents,
     get_family_document_and_context,
     get_slugged_objects,
+)
+from app.schemas.document import (
+    FamilyAndDocumentsResponse,
+    FamilyDocumentWithContextResponse,
 )
 
 _LOGGER = logging.getLogger(__file__)
