+++ app/api/api_v1/routers/lookups/config.py
@@ -1,9 +1,9 @@
 from fastapi import Depends, Request
 
 from app.api.api_v1.routers.lookups.router import lookups_router
-from app.api.api_v1.schemas.metadata import ApplicationConfig
 from app.clients.db.session import get_db
 from app.repository.lookups import get_config
+from app.schemas.metadata import ApplicationConfig
 
 
 @lookups_router.get("/config", response_model=ApplicationConfig)
