+++ app/core/browse.py
@@ -10,13 +10,13 @@ from db_client.models.organisation import Organisation
 from pydantic import BaseModel
 from sqlalchemy.orm import Session
 
-from app.api.api_v1.schemas.search import (
+from app.repository.geography import get_geo_subquery
+from app.schemas.search import (
     SearchResponse,
     SearchResponseFamily,
     SortField,
     SortOrder,
 )
-from app.repository.geography import get_geo_subquery
 
 _LOGGER = getLogger(__name__)
