+++ app/repository/lookups.py
@@ -7,10 +7,10 @@ from db_client.models.document.physical_document import Language
 from sqlalchemy.exc import MultipleResultsFound
 from sqlalchemy.orm import Session
 
-from app.api.api_v1.schemas.metadata import ApplicationConfig
 from app.core.util import tree_table_to_json
 from app.core.validation import IMPORT_ID_MATCHER
 from app.repository.organisation import get_all_organisations, get_organisation_config
+from app.schemas.metadata import ApplicationConfig
 
 _LOGGER = logging.getLogger(__name__)
