+++ app/service/pipeline.py
@@ -3,9 +3,9 @@ from typing import Any, Sequence
 
 from fastapi import Depends
 
-from app.api.api_v1.schemas.document import DocumentParserInput
 from app.clients.db.session import get_db
 from app.repository.pipeline import generate_pipeline_ingest_input
+from app.schemas.document import DocumentParserInput
 
 _LOGGER = logging.getLogger(__name__)
