+++ tests/non_search/core/validation/test_util.py
@@ -5,10 +5,10 @@ from unittest import mock
 
 import pytest
 
-from app.api.api_v1.schemas.document import DocumentParserInput
 from app.config import PIPELINE_BUCKET
 from app.core.validation import IMPORT_ID_MATCHER
 from app.core.validation.util import _flatten_maybe_tree, write_documents_to_s3
+from app.schemas.document import DocumentParserInput
 
 NOT_A_TREE_1 = [{"name": 1}, {"name": 2}, {"name": 3}]
 NOT_A_TREE_2 = [
